feat: placed in separate modules basic/zip
