+++ src/basic/rm.js
@@ -1,3 +1,5 @@
+import { unlink } from "fs";
+import { join } from "path";
 
 export default function handleRmCommand(command, currentPath) {
   if (!command) {
